requirements.txt
tweepy==4.14.0
python-dotenv==1.0.0
feedparser==6.0.10
requests==2.31.0
schedule==1.2.0
nltk==3.8.1
pandas==2.0.3
matplotlib==3.7.2
textblob==0.17.1
pytest==7.4.0 